Early IPv6 support
This introduces the version 11 for the host schema. Version 11 stores
the IP addresses as a string containing their canonical representation
(A.B.C.D or AAAA:BBBB:CCCC:DDDD::1).

In MongoDB, the addresses are internally stored as a string
corresponding to the hexadecimal representation of the 128-bit integer
corresponding to the IPv6 address. IPv4 addresses are represented as
"IPv4-Mapped IPv6 Addresses", as defined in RFC 4291 section 2.5.5.2
(see <https://tools.ietf.org/html/rfc4291#section-2.5.5.2>), i.e.,
::ffff:10.0.0.1.

The reasons for this choice are the impossibility to store (and hence,
index) unsigned 128-bit integers in MongoDB. Another approach, storing
the 128-bit integer as a binary blob has been considered but the lack
of operators to manipulate binary blobs in the aggregation framework
would have made some analysis operations impossible to do "in
database".

In PostgreSQL, the INET type already supports IPv6 addresses.

In SQLite, the addresses are internally stored as a 16-byte binary blob
corresponding to the IPv6 address.