Travis-CI bug: disable PostgreSQL w/ Python 3.5 tests
The command "ivre ipdata --download" segfaults in Travis-CI env (Bionic)
with Python 3.5 and PostgreSQL, for some reason that does not look like
related to IVRE's code (and that I have not been able to reproduce outside
of Travis-CI workers).

I don't have time to investigate. PostgreSQL is still tested with Python
2.7, 3.4 (under Xenial) and 3.6 to 3.8.