refactor(ci): replace unmaintained PSQL action with built-in `services` setup
This commit replaces unmaintained `harmon758/postgresql-action` action with built-in GHA `services` to pop a PSQL container for testing purposes.
As a bonus, it also brings container healthcheck to make sure we wait for this container is ready before proceeding to tests.

Taken from: https://docs.github.com/en/actions/use-cases-and-examples/using-containerized-services/creating-postgresql-service-containers